backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-phonenumber-field==8.4.0
djangorestframework==3.17.0
phonenumbers==9.0.26
sqlparse==0.5.5

python-dotenv~=1.2.2